Task1_/requirements.txt
pillow 
pytesseract
ollama
